First commit.
Includes a first initial working version of an SFU model, which receives
data from one end and pushes it towards the other connected ends.

Arch hasn't been fleshed out yet, but here's how things are layed
out:
- convo: the conference module holds the structure and logic for each
  member, and the conferences itselves;
- protos: the protocols module holds each of the implemented protocols
  that users can iterface with in order to create conferences / members
  and, eventually, place calls;
- sdp.rs: a very simple and incomplete (for now) SDP parser which turns
  a string that represents an SDP into a struct and vice versa;
- main.rs: where it all start and the rest of the components are
  initialised;
- test.rs and test_receiver.rs: tests written to test the gateway
  functionality, end-to-end. Ideally they will be moved to a tests/
  folder.

The Conference struct in convo/convo.rs defines the structure of a
conference, which can have multiple members, an Id, and an SDP (which
defines the common SDP all the members need to agree on). In turn, each
member (struct Member in convo/member.rs) also has an SDP and an
RtpSession.

The handlers trait in the protos module defines the common operations
all protocols needs to implement; For now, that's only the start_server
operation.  The protocols currently implemented, tcpserver and
httpserver (which relies on nickel), are the only ways to interact with
the gateway - the tcpserver being way less granular than the httpserver
protocol, which provides REST endpoints.

sdp.rs provides the SessionDescription struct, which is a representation
of an SDP. The method from_sdp allows to turn a raw SDP string / byte
buffer to a ParseResult (which may contain the errors eventually found
during the parsing, but also the SessionDescription itself); On the
other hand, SessionDescription also implements the ToString trait,
allowing to easily convert it into a string.

Finally, as briefly mentioned before, simple tests using
test_receiver.rs and test.rs (which rely on rir's implementation of RTP)
have been run to guarantee that a WAV file, encoded in OPUS and sent by
test.rs, can be received and decoded correctly on the test_receiver.rs
side - i.e. the original file is duplicated at the end.