Generalise the concept of members.
Members have been only a part of a conference, not even having an Id.
However, in order to be able to refer to them from a global point of
view (for example, in the /members REST endpoint), not only they need to
have an Id, but they also need to be on a globally available pool, just
like conferences.

The convo/member.rs has been changed to reflect the new design, being
now very similar to the conference, with a global HashMap that holds all
the members created in the platform.

A particular difference is that each member is wrapped around a
read-write lock (instead of just an Arc), since there are operations we
need to guarantee that are only accessed by one thread at a time.

The Conference struct now also holds a HashMap of members, identified by
their Id, instead of just refering through a vector of members - so that
one can easily find a member inside a conference.