Finish concept of member in httpserver.
Although the functions for dealing with members and the MemberResponse
and MemberPost were already present, they weren't yet being used, since
there was no concept of "identifiable member" internally, and thus some
work was in need to link conference to members at the http protocol
level.

The get_member function was changed in order to get the member at a
global level ("/member/:memberid"), while a new function,
get_conference_member, was introduced in order to get the member at a
conference level ("/convo/:convoid/member/:memberid").

MemberResponse was also modified in order to support the SDP.