Change test examples to use HTTP.
The test examples, test.rs and test_receiver.rs, were still based on the
TCP interface.

They're now using using the main interface, the HTTP one, to be easier
to check and grasp for an working example;
The tests use the HTTP client provided by hyper, and depend on the
MemberPost, MemberResponse, ConferencePost and ConferenceResponse
defined in protos/httpserver.rs.

Some cleanup needs to be done, in order to move the examples into a test
folder, and even bring back the TCP examples back to life, to have
working examples on that interface as well.