Move integration tests to proper folder.
Best practises dictate that integration tests should be in a tests/
folder at the root of the project - hence moved the test_receiver.rs and
test,rs to test_receive.rs and test_send.rs, respectively.

As part of this changes, and in order to be able to run the integration
tests, a new lib.rs file was also added - the file itself declares the
exported moduled so the tests can be used.

The Cargo.toml was also cleaned up in order to get rid of the binaries
we were creating proposedly for the test files.