Improve the concept of media in the SDP.
The SDP support in sdp.rs only provides support for one media and
attribute in the SessionDescription structure.

In order to comply with the RFC#4566 multiple attributes can be hold at
the session level (anything before the first media), and the same at the
media level (the Media struct).  For this, the from_sdp method was
modified in order to push the parsed media and attributes to the
appropriate places - which might be the attributes in the
SessionDescription or the attributes in the MediaDescription. The latter
happens after reading the first Media, i.e.  all attributes after that
belong to that media session (m=), until the next media session, and so
on.

However, the work on convo.rs and member.rs and even the ToString's
implementation of the SessionDEscription struct are hard-coded to red
the first media only.