Clean SessionDescription's string transformation.
Instead of reading the values from each of SessionDescription's member
fields (and respective values, and so on), and trying to decode how to
print each of the values, it now calls the to_string method where
appropriate, to get the string representation of a specific value, like
NetType.

In order to achieve this the ToString trait was implemented for the
following structs - AddrType, NetType, MediDescription, Attr, Connection
and Origin.