Improve SDP's attribute parsing.
Although parsing an attribute is taking into account both property and
value attributes, it is still too generic in order to be able to then
use it properly (when comparing, for example, during negotiation).

To make it more usable the Attr struct is now an enum instead, which
contains several variants, which may be parameterized or not.

The Attr now implements the AttrFromStr trait which converts either a
property attribute or a value attribute into the appropriate enum.  The
ToStr implementation was also modified to convert the enums into
strings.

The parse_attr function was modified to call the from_str function
implemented by the Attr, to parse the attribute appropriately.