Fix value attribute parsing.
Parsing value attributes (i.e. of format "rtpmap:8 PCMA/8000") was
failing; In the parse_attr function, the string passed was being split
by spaces (like in most other parsing functions), and not the more
appropriate colon character.

Thus parse_attr was modified to parse the passed string based on a
colon, which indicates if an attribute is a property attribute (in which
case it shouldn't have a colon), or a value attribute (in which case it
should have a colon, and needs extra parsing).