Add support for rtpmap and fmtp attributes.
To parse the rtpmap and fmtp value attributes support was added by
adding two new variants to the Attr enum, RtpMap and FmtP - the former
for rtpmap and the latter for fmtp.

These new variants work as PTime, and thus use RtpMapValue and FmtpValue
structs as (unnamed) data.
The FromStr and ToString implementations of each of those respective
structures is then responsible of parsing the value of the respective
attribute, from a string, or transform it into a string.

Lastly, they also needed to be added to the ToString and FromStr
implementations for Attr, in order to be considered when parsing an
attribute.