Improve SDP negotiation based on rfc#3264.
SDP negotiation was only a simple clone of the original offer, which
means it wasn't respecting much of what's defined in rfc#3264.

This patch introduces the concept of media negotiating, to compare SDPs
and generate an answer based on the offer.

The negotiate_media function was renamed to negotiate_media_stream, in
order to reflect that it only negotiates one stream at a time.
Although still not yet complete, the function now compares the media
types and protocols between both SDPs provided (the common one and the
offer), and then modifies the offer in order to turn it into a response,
based on the diffferent attributes:
- The direction attributes (sendrecv, sendonly, ...) are reversed only,
  for now;
- ptime is only copied;
- The format types of the offer are compared to the common SDP and only
  those types and respective rtpmap's are kept on the offer.

The ngotiate_with function is the one that goes through the the media in
each SDP and compares them, which then gets a match and an answer or
sets the port of that stream to 0 (effectively rejecting it).

Also updated test_send to add a new format type to the media, opus, for
tetsing purposes.

Further improvements will include to respect RFC#3264 more thoroughly
and also improve the negotiate_media_stream function to be more readable
(in a Rust way).