gh-3: ICE Lite initial support.
Added support for a few ICE attributes in the SDP parsing and
serialising:
- Candidate;
- IceUfrag;
- IcePwd;
- IceMismatch;
- IceLite.
At the same time the negotiate_media_stream and negotiate_with routines
were reorganised to take ICE parsing into account.

The convo and member modules were also adapted to the new use a Mutex
around a Member, instead of a Read / Write lock.
The Member contains an Ice instance while the init_audio was modified to
hardcode the local IP - needs to be changed to use the Candidates.

The new ice.rs module is still in early stages but already enables to
gather the Host candidates required by ICE Lite.

Also added to Cargo.toml the dependencies needed, rustun (for STUN),
ifaces and fibers (to use rustun).

Note that this patch is still in testing phase, and there's still work
to do in order to have a complete implementation - for example a patch
to Rustun will need to be pushed.