gh-3: Organise candidates by component id.
Each gathered candidate is now stored for a specific component id in a
stream, both locals and remotes. Thus, when gathering the candidates,
instead of providing the number of total components for the stream, the
component itself needs to be provided.

The logic between finding a local IPv4 address for the candidates as
also been split from gather_candidates into a separate function,
get_ipv4_address;