gh-3: Clean up ice.rs and gather_candidates.
The gather_candidates method in the Agent was cleaned up in order to
improve the logic around adding a new candidate to a component. It now
also sets up the stun server on the UdpSocket before returning the
candidate, by using the setup_stun_server which has been uncommented
again.

Compilation and borrow check errors were also cleaned in the ice.rs
module.

Also updated Cargo to a locally modified version of fibers-rs.