gh-3: Integrate session_negotiation.rs and ice.rs.
The session_negotiation.rs module has been modified in order to
integrate with the new ICE API. Thus, three main routines were defined:
- process_offer: Goes through the media streams, and for each adds a new
  stream on the instantiated ICE agent and triggers candidates
retrieval;
- negotiate_with_base_sdp: Compares two SDPs, the offer with the base
  SDP set at the convo level;
- process_answer: Process the SDP answer, gets the candidates that were
  ipreviously gathered for the offer and injects them into the answer.

The Session now stores a base, answer and offer  SDPs and a mapping
which maps the positions of each stream Ids in the SDP. I.e. if the
stream in the SDP comes first, the correspondent stream Id will be first
in the queue as well.

The ICE agent now has a get_stream_candidates which retrieves the
candidates that were gathered for a specific component of a specific
stream.