gh-3: Fix compilation and borrow checker errors.
The member.rs module has also been fixed to use the
session_negotiation.rs module, instead of the ice.rs module.

In the session_neogitation.rs module, Session is now public and all
methods were modified to accept mutable references of self, since the
calls to the Agent to add_stream and gather_candidates would also
require so.
The order of operations in negotiate_with_base_sdp were modified so an
unnecessary clone of the base SDP isn't necessary.

In the ice.rs module both RTP_COMPONENT_ID and RTCP_COMPONENT_ID
constants were made public, to be exported and used outside.

Finally, the sdp.rs module was also modified to get rid of the logic to
add the Candidate attribute during the negotiation.