gh-3: Use Member's negotiate_session in convo.rs.
A new method, negotiate_session, was added to Member in the member.rs
module, which uses the stores offer SDP and compares it to the SDP
passed, using the Session already instantiated.
Another method has also been introduced, get_session_answer, in order
to be able to retrieve the answer when the SDP has been negotiated.

When adding a member to a convo, the member's provided SDP offer still
needs to be negotiate with convo's base SDP. This is now done by using
the negotiate_session in Member, instead of using the one in sdp.rs
(which is still used by the session_negotiation.rs module).

In the Ice.rs module, Agent's method get_stream_candidates was modified
to return an Option instead, so it is clear if there are candidates or
not.