gh-3: Initalise media sessions for each candidate.
In the session_negotiation.rs module, while executing process_answer,
the media sessions capable of dealing with STUN need to be initialised
for each candidate gathered - as specified in rfc5245, the agent needs
to be prepared to receive STUN and media packets.

Moved the init_media_session and new_rtp_session from the member.rs
module to the session_negotiation.rs module, where the former is a
method of Session.

Thus, in process_answer, For each candidate found, init_media_session is
called in order to set a new RTP session.