gh-3: Pair Candidates and support callbacks.
A new method has been introduced to trigger the state change of a stream
(to completed), in case the valid list has a candidate for each
component.

The concept of a handler has been added to the Agent in ice.rs. There's
a Handler trait which each user who wants to receive callbacks regarding
certain events (like state updates) must implement and pass onto in the
start() method, after instantiating the Agent.

Implementing incomplete pair_candidates method which gathers the local /
offer candidates and compares them in order to form the valid list,
which is a list of PairCandidates.
pair_candidates then calls trigger_state_change to check if the state
needs to be updated and delivers the event to the registered callback.

The start method was also updated to start a thread which periodically
checks the overall state of all streams and updates overall ice state of
the agent to "Completed" if all streams are "Completed".

Finally the session_negotiation.rs module was updated to iterate and
push each candidate on each stream of the SDP to the agent, when
processing the offer, and implement the trait Hanler from ice.rs.

Updated Cargo to import the timer and time crates.