Remove uneeded init() method from Convo.
The init() method in Convo is redundant, since it is there only to start
the processing engine, by calling process_engine(). This has now been
replaced by a call to process_engine() when the first member is added.

Also, remove init() calls from the proto/ implementations. These were
the only dependencies so far.