Fix SDP negotiation against existing Convo SDP.
When comparing an offer SDP with an already existing Convo SDP (one that
has been bound to a convo), negotiation was failing because of two
factors:
1. Candidates from the offer weren't being removed;
2. The offered "ice-ufrag" and "ice-pwds" were being delivered back in
the answer.

This fixes it by deleting the offered candidates and ice parameters, and
injecting new ones (the local candidates are injected, as it was already
happening).