Give interior mutability to Session.
Previously, when members were to use an instance Session, they would
need to be aware of mutability, as some methods would expect mutable
instances. This would mean that the session member in the MemberSession
struct would have a Mutex around the Session. This could be potentially
create high contention, since during write and read of RTP packets the
member would have to lock the entire session (and thus coudln't write
and read at the same time).

To change that. the Session struct now provides interior mutability, and
thus the a Member can now use a Session directly as it will lock and
unlock were necessary internally. This way we can read and write at the
same time.

The media_sessions member in the Ssssion struct was also changed from a
Mutex to a RwLock, to avoid uneeded contention during reads.