Depend on the official opus-rs.
opus-rs doesn't implement the Send trait, meaning an instance of the
encoder / decoder can't be sent between threads. Previously, locally, a
version of opus-rs was modified to unsafely implement the Send trait.
However, this commit now reverts that and makes hibrido now depend on
the official version.

For this, each instance of the decoder / encoder is now only
instantiated inside of each worker thread, which are now also abstracted
by a MemberWorker struct that spawns the worker threads.