Lose lazy_static dependency.
Lazy_static was being used so that we could have a global static hash
map to store each conference instance. This was not ideal, having a
global static, and because of its nature, lazy_static could also
introduce, sometimes, some unclarity while understanding an issue.

To lose the dependency, a hashmap of Conferences is now initialiased
during startup and passed as a dependency where it is required. At this
point to the protos/ implementations.